backend/pyproject.toml
[project]
name = "Training-FastAPI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.23.0",
    "pre-commit>=3.0.0",
]
